feat(class/index.js) wrapper remove, file renamed
